ScriptsController#create now creates DataCollector
when create a Script now you can choose the
targets that it will runs. BUT there is a bug
when there are errors and it render the 'new'
action, so i'm commiting now until I drink
some coffee and continue to try figure out
how to find the problem... well.. this
branch is broken, don't use it!